Fix: (#24, #25) JobOffer: Decoupling logic, Unit Tests and Refactoring

+++ src/main/java/com/alumniportal/unmsm/service/impl/JobOfferServiceImpl.java
@@ -1,6 +1,9 @@
 package com.alumniportal.unmsm.service.impl;
 
+import com.alumniportal.unmsm.dto.RequestDTO.JobOfferRequestDTO;
 import com.alumniportal.unmsm.dto.ResponseDTO.JobOfferResponseDTO;
+import com.alumniportal.unmsm.exception.AppException;
+import com.alumniportal.unmsm.mapper.JobOfferMapper;
 import com.alumniportal.unmsm.model.Company;
 import com.alumniportal.unmsm.model.JobOffer;
 import com.alumniportal.unmsm.persistence.interfaces.ICompanyDAO;
@@ -24,23 +27,24 @@ public class JobOfferServiceImpl implements IJobOfferService {
 
     private final ICompanyDAO companyDAO;
 
-    private final ModelMapper modelMapper;
+    private final JobOfferMapper jobOfferMapper;
 
     @Override
     public List<JobOfferResponseDTO> findAll() {
-        return jobOfferDAO.findAll()
-                .stream()
-                .map(jobOffer -> modelMapper.map(jobOffer, JobOfferResponseDTO.class))
-                .toList();
+        List<JobOffer> jobOfferList = jobOfferDAO.findAll();
+        if (jobOfferList.isEmpty()) {
+            throw new AppException("No job offers found!", "NOT_FOUND");
+        }
+        return jobOfferMapper.entityListToDTOList(jobOfferList);
     }
 
     @Override
     public JobOfferResponseDTO findById(Long id) {
         JobOffer jobOffer = jobOfferDAO.findById(id);
         if (jobOffer == null) {
-            return null;
+            throw new AppException("Job offer not found!", "NOT_FOUND");
         }
-        return modelMapper.map(jobOffer, JobOfferResponseDTO.class);
+        return jobOfferMapper.entityToDTO(jobOffer);
     }
 
     @Override
@@ -50,36 +54,42 @@ public class JobOfferServiceImpl implements IJobOfferService {
 
     @Override
     public void deleteById(Long id) {
+        JobOffer jobOffer = jobOfferDAO.findById(id);
+        if (jobOffer == null) {
+            throw new AppException("Job offer not found!", "NOT_FOUND");
+        }
         jobOfferDAO.deleteById(id);
     }
 
     @Override
     public List<JobOfferResponseDTO> findJobOffersByCompanyId(Long id) {
-        return jobOfferDAO.findJobOffersByCompanyId(id)
-                .stream()
-                .map(jobOffer -> modelMapper.map(jobOffer, JobOfferResponseDTO.class))
-                .toList();
+        List<JobOffer> jobOffersByCompanyId = jobOfferDAO.findJobOffersByCompanyId(id);
+        if (jobOffersByCompanyId.isEmpty()) {
+            throw new AppException("No job offers found for this company!", "NOT_FOUND");
+        }
+        return jobOfferMapper.entityListToDTOList(jobOffersByCompanyId);
     }
 
     @Override
-    public void saveJobOffer(JobOffer jobOffer, Long companyId) {
+    public void saveJobOffer(JobOfferRequestDTO jobOfferRequestDTO, Long companyId) {
         Company company = companyDAO.findById(companyId);
         if (company == null) {
-            throw new RuntimeException("Company not found!");
+            throw new AppException("Company not found!", "NOT_FOUND");
         }
+
+        JobOffer jobOffer = jobOfferMapper.requestDtoToEntity(jobOfferRequestDTO);
+
         jobOffer.setCompany(company);
         jobOffer.setCreatedAt(LocalDate.now());
         jobOfferDAO.save(jobOffer);
 
-        company.getJobOfferList().add(jobOffer);
-        companyDAO.save(company);
     }
 
     @Override
     public void updateJobOffer(Long id, Map<String, Object> fields) {
         JobOffer jobOffer = jobOfferDAO.findById(id);
         if (jobOffer == null) {
-            throw new RuntimeException("Error: jobOffer not found!");
+            throw new AppException("Job offer not found!", "NOT_FOUND");
         }
         fields.forEach((k, v) -> {
             Field field = ReflectionUtils.findField(JobOffer.class, k);
